fix community contract tests

+++ src/database/migrations/z1604506690-create-communityContract.js
@@ -32,7 +32,7 @@ module.exports = {
                 allowNull: false,
             },
             decreaseStep: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.DECIMAL(22), // max 9,999 - plus 18 decimals
                 allowNull: false,
                 defaultValue: 0,
             },
